Use SmartCSV for daily import

+++ app/services/import_offenders.rb
@@ -2,11 +2,13 @@ module ImportOffenders
   module_function
 
   def call(import)
-    ParseOffenders.call(import.offenders_file.read)
-    import.update_attribute(:status, :successful)
-    Import.where('id != ?', import.id).destroy_all
-  rescue StandardError => e
-    import&.update_attribute(:status, :failed)
-    NotificationMailer.import_failed(import, e.to_s).deliver_now
+    errors = ParseOffenders.call(import.offenders_file.file.file)
+    if errors.any?
+      import.update_attribute(:status, :failed)
+      NotificationMailer.import_failed(import, errors.join("\n\n")).deliver_now
+    else
+      import.update_attribute(:status, :successful)
+      Import.where('id != ?', import.id).destroy_all
+    end
   end
 end
